refactor(ui): Convert UI components to constexpr observer pattern
Removes the runtime `etl::observer` pattern from `AnalogControl`, `Button`, `Drumpad`, and `Keypad_HC138` and replaces it with the compile-time `musin::observer` pattern.

This change achieves the following:
- Converts the UI component classes into templates that accept a parameter pack of observer pointers (`auto*... Observers`).
- Eliminates runtime overhead (RAM and CPU) for observer notifications by resolving them at compile time.
- Moves all implementation code from `.cpp` files into the headers, preparing for template instantiation.
- Deletes the now-redundant `.cpp` source files.
- Enhances `musin::observer` with C++20 concepts to ensure type safety for observers.

This is the first step in a larger refactoring. The instantiation of these components will be updated in subsequent commits.
